Add ErrorBoundary to DashboardView
JIRA: RAIL-2755
