+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/index.ts
@@ -1,0 +1,4 @@
+// (C) 2020 GoodData Corporation
+import { DashboardViewErrorBoundary } from "./DashboardViewErrorBoundary";
+export { DashboardViewErrorBoundary as DashboardView };
+export { IDashboardViewProps } from "./types";
