+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/types.ts
@@ -1,21 +1,14 @@
 // (C) 2020 GoodData Corporation
-import React, { useEffect } from "react";
-import { IAnalyticalBackend, IDashboard, ITheme, IWidgetAlert } from "@gooddata/sdk-backend-spi";
-import { IFilter, ObjRef } from "@gooddata/sdk-model";
+import { IAnalyticalBackend, ITheme, IDashboard, IWidgetAlert } from "@gooddata/sdk-backend-spi";
+import { ObjRef, IFilter } from "@gooddata/sdk-model";
 import {
     IDrillableItem,
-    IErrorProps,
     IHeaderPredicate,
-    ILoadingProps,
-    ErrorComponent as DefaultError,
-    LoadingComponent as DefaultLoading,
-    OnError,
     OnFiredDrillEvent,
+    IErrorProps,
+    ILoadingProps,
+    OnError,
 } from "@gooddata/sdk-ui";
-import { ThemeProvider, useThemeIsLoading } from "@gooddata/sdk-ui-theme-provider";
-import { useDashboard } from "../useDashboard";
-import { DashboardRenderer } from "./DashboardRenderer";
-import { useDashboardAlerts } from "../useDashboardAlerts";
 
 export interface IDashboardViewProps {
     /**
@@ -94,75 +87,3 @@ export interface IDashboardViewProps {
      */
     onError?: OnError;
 }
-
-export const DashboardView: React.FC<IDashboardViewProps> = ({
-    dashboard,
-    filters,
-    theme,
-    backend,
-    workspace,
-    onDrill,
-    drillableItems,
-    onError,
-    onDashboardLoaded,
-    ErrorComponent = DefaultError,
-    LoadingComponent = DefaultLoading,
-}) => {
-    const { error: dashboardError, result: dashboardData, status: dashboardStatus } = useDashboard({
-        dashboard,
-        onError,
-    });
-
-    const { error: alertsError, result: alertsData, status: alertsStatus } = useDashboardAlerts({
-        dashboard,
-        onError,
-    });
-
-    const isThemeLoading = useThemeIsLoading();
-    const hasThemeProvider = isThemeLoading !== undefined;
-
-    useEffect(() => {
-        if (alertsData && dashboardData) {
-            onDashboardLoaded?.({
-                alerts: alertsData,
-                dashboard: dashboardData,
-            });
-        }
-    }, [onDashboardLoaded, alertsData, dashboardData]);
-
-    const statuses = [dashboardStatus, alertsStatus];
-
-    if (statuses.includes("loading") || statuses.includes("pending")) {
-        return <LoadingComponent />;
-    }
-
-    if (dashboardStatus === "error") {
-        return <ErrorComponent message={dashboardError.message} />;
-    }
-
-    if (alertsStatus === "error") {
-        return <ErrorComponent message={alertsError.message} />;
-    }
-
-    const dashboardRender = (
-        <DashboardRenderer
-            dashboard={dashboardData}
-            alerts={alertsData}
-            filters={filters}
-            onDrill={onDrill}
-            drillableItems={drillableItems}
-            ErrorComponent={ErrorComponent}
-            LoadingComponent={LoadingComponent}
-        />
-    );
-
-    if (!hasThemeProvider) {
-        return (
-            <ThemeProvider theme={theme} backend={backend} workspace={workspace}>
-                {dashboardRender}
-            </ThemeProvider>
-        );
-    }
-
-    return dashboardRender;
-};
